Split ComActivator into separate files for supported and not supported
